pyproject.toml
[project]
name = "agent-example"
version = "0.1.0"
description = "A deep agent for various tasks"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dashscope>=1.25.13",
    "ddgs>=9.0.0",
    "deepagents>=0.4.11",
    "duckduckgo-search>=8.0.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "python-dotenv>=1.0.0",
]
